backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
yt-dlp==2025.10.22
google-api-python-client==2.187.0
openai==2.7.1
requests==2.32.5
python-dotenv==1.0.0
firebase-admin==7.1.0
gunicorn==21.2.0
pytest==7.4.3
flake8==6.1.0
